move utility fns into module
