+++ gql/count_pull_requests.graphql
@@ -1,0 +1,5 @@
+query CountPullRequests($query_string: String!) {
+  search(query: $query_string, type: ISSUE, last: 1) {
+    issueCount
+  }
+}
