--- gql/query_search.graphql
@@ -1,5 +1,0 @@
-query QuerySearch($query_string: String!) {
-    search(query:$query_string, type:ISSUE,last:1) {
-        issueCount
-    }
-}
