+++ src/metrics/all_repos.rs
@@ -1,5 +1,6 @@
 use super::Graphql;
 use anyhow::Error;
+use chrono::{Duration, Utc};
 use fehler::throws;
 use graphql_client::GraphQLQuery;
 
@@ -12,7 +13,7 @@ use graphql_client::GraphQLQuery;
 struct OrgRepos;
 
 #[throws]
-pub async fn all_repos_graphql(graphql: &Graphql, org: &str) -> Vec<String> {
+pub(super) async fn all_repos(graphql: &Graphql, org: &str) -> Vec<String> {
     let org_name = format!("{}", org);
     let mut repos: Vec<String> = vec![];
     let mut after_cursor = None;
@@ -52,4 +53,47 @@ pub async fn all_repos_graphql(graphql: &Graphql, org: &str) -> Vec<String> {
     }
 
     repos
+}
+
+#[derive(GraphQLQuery)]
+#[graphql(
+    schema_path = "gql/schema.docs.graphql",
+    query_path = "gql/count_pull_requests.graphql",
+    response_derives = "Serialize,Debug"
+)]
+struct CountPullRequests;
+
+/// count the number of pull requests created in the given time period for the given repository within the given GitHub organization
+///
+/// # Arguments
+/// - `org_name` — The name of the github organization that owns the specified repository
+/// - `repo_name` — The name of the repository to count pull requests for. **Note:** repository should exist within the `org_name` Github Organization
+/// - `time_period` — The relevant time period to search within
+#[throws]
+pub(super) async fn count_pull_requests(
+    graphql: &Graphql,
+    org_name: &str,
+    repo_name: &str,
+    time_period: Duration,
+) -> usize {
+    // get date string to match GitHub's PR query format for `created` field
+    // i.e., "2021-05-18UTC" turns into "2021-05-18"
+    let date_str = chrono::NaiveDate::parse_from_str(
+        &format!("{}", (Utc::now() - time_period).date())[..],
+        "%FUTC",
+    )
+    .unwrap();
+
+    let query_string = format!(
+        r#"repo:{}/{} is:pr created:>{}"#,
+        org_name, repo_name, date_str,
+    );
+
+    let response = graphql
+        .query(CountPullRequests)
+        .execute(count_pull_requests::Variables { query_string })
+        .await?;
+    let response_data = response.data.expect("missing response data");
+    let count = response_data.search.issue_count;
+    count as usize
 }
