+++ src/metrics/list_repos.rs
@@ -1,11 +1,8 @@
-use anyhow::Error;
 use async_trait::async_trait;
-use chrono::{Duration, Utc};
-use fehler::throws;
-use graphql_client::GraphQLQuery;
+use chrono::Duration;
 use tokio::sync::mpsc::Sender;
 
-use super::{Graphql, Producer};
+use super::{all_repos, Graphql, Producer};
 
 #[derive(Debug)]
 pub struct ListReposForOrg {
@@ -31,10 +28,10 @@ impl Producer for ListReposForOrg {
     }
 
     async fn producer_task(self, tx: Sender<Vec<String>>) -> Result<(), anyhow::Error> {
-        let repos: Vec<String> = super::all_repos_graphql(&self.graphql, &self.org_name).await?;
+        let repos: Vec<String> = all_repos::all_repos(&self.graphql, &self.org_name).await?;
 
         for repo in &repos {
-            let count_prs = count_pull_requests_graphql(
+            let count_prs = all_repos::count_pull_requests(
                 &self.graphql,
                 &self.org_name,
                 &repo,
@@ -47,47 +44,4 @@ impl Producer for ListReposForOrg {
 
         Ok(())
     }
-}
-
-#[derive(GraphQLQuery)]
-#[graphql(
-    schema_path = "gql/schema.docs.graphql",
-    query_path = "gql/query_search.graphql",
-    response_derives = "Serialize,Debug"
-)]
-struct QuerySearch;
-
-/// count the number of pull requests created in the given time period for the given repository within the given GitHub organization
-///
-/// # Arguments
-/// - `org_name` — The name of the github organization that owns the specified repository
-/// - `repo_name` — The name of the repository to count pull requests for. **Note:** repository should exist within the `org_name` Github Organization
-/// - `time_period` — The relevant time period to search within
-#[throws]
-async fn count_pull_requests_graphql(
-    graphql: &Graphql,
-    org_name: &str,
-    repo_name: &str,
-    time_period: Duration,
-) -> usize {
-    // get date string to match GitHub's PR query format for `created` field
-    // i.e., "2021-05-18UTC" turns into "2021-05-18"
-    let date_str = chrono::NaiveDate::parse_from_str(
-        &format!("{}", (Utc::now() - time_period).date())[..],
-        "%FUTC",
-    )
-    .unwrap();
-
-    let query_string = format!(
-        r#"repo:{}/{} is:pr created:>{}"#,
-        org_name, repo_name, date_str,
-    );
-
-    let response = graphql
-        .query(QuerySearch)
-        .execute(query_search::Variables { query_string })
-        .await?;
-    let response_data = response.data.expect("missing response data");
-    let count = response_data.search.issue_count;
-    count as usize
 }
